+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.8.0-dev
   **NEW FEATURES**
   - [#174](https://github.com/Datatamer/unify-client-python/issues/174) Get and create taxonomy categories
+  - [#182](https://github.com/Datatamer/unify-client-python/issues/182) Add the ability to refresh estimated pair counts.
 
 ## 0.7.0
   **BREAKING CHANGES**